+++ Examples/XestiMonitorsDemo-iOS/XestiMonitorsDemo/MasterViewController.swift
@@ -11,7 +11,7 @@ import UIKit
 
 public class MasterViewController: UITableViewController {
     override public func viewWillAppear(_ animated: Bool) {
-        self.clearsSelectionOnViewWillAppear = self.splitViewController?.isCollapsed ?? false
+        clearsSelectionOnViewWillAppear = splitViewController?.isCollapsed ?? false
 
         super.viewWillAppear(animated)
     }