+++ Sources/Core/Other/KeyboardMonitor.swift
@@ -249,32 +249,6 @@ public class KeyboardMonitor: BaseNotificationMonitor {
             }
         }
     }
-
-    // MARK: Deprecated
-
-    ///
-    /// Initializes a new `KeyboardMonitor`.
-    ///
-    /// - Parameters:
-    ///   - queue:      The operation queue on which the handler executes.
-    ///                 By default, the main operation queue is used.
-    ///   - options:    The options that specify which events to monitor.
-    ///                 By default, all events are monitored.
-    ///   - handler:    The handler to call when the visibility of the
-    ///                 keyboard or the frame of the keyboard changes or is
-    ///                 about to change.
-    ///
-    /// - Warning:  Deprecated. Use `init(options:queue:handler)` instead.
-    ///
-    @available(*, deprecated, message: "Use `init(options:queue:handler)` instead.")
-    public init(queue: OperationQueue = .main,
-                options: Options = .all,
-                handler: @escaping (Event) -> Void) {
-        self.handler = handler
-        self.options = options
-
-        super.init(queue: queue)
-    }
 }
 
 #endif