+++ Tests/Screen/ScreenBrightnessMonitorTests.swift
@@ -1,6 +1,6 @@
 //
 //  ScreenBrightnessMonitorTests.swift
-//  XestiMonitors
+//  XestiMonitorsTests
 //
 //  Created by Paul Nyondo on 2018-03-25.
 //